# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handlers.CampaignPerformanceListResponse:
    properties:
      campaigns:
        items:
          $ref: '#/definitions/internal_handlers.CampaignPerformanceResponse'
        type: array
      count:
        example: 3
        type: integer
    type: object
  internal_handlers.CampaignPerformanceResponse:
    properties:
      campaign_id:
        type: string
      campaign_name:
        type: string
      ctr:
        type: number
      spent_cents:
        type: integer
      status:
        type: string
      total_clicks:
        type: integer
      total_impressions:
        type: integer
    type: object
...
  internal_handlers.DashboardOverviewResponse:
    properties:
      active_campaigns_count:
        example: 5
        type: integer
      budget_utilization:
        example: 42
        type: number
      ctr:
        example: 0
        type: number
      total_budget_cents:
        example: 100000
        type: integer
      total_clicks:
        example: 0
        type: integer
      total_impressions:
        example: 0
        type: integer
      total_spent_cents:
        example: 42000
        type: integer
    type: object
...
  internal_handlers.ErrorDetail:
    properties:
      code:
        example: STATION_NOT_FOUND
        type: string
      field:
        example: email (only present on validation errors)
        type: string
      message:
        example: station not found
        type: string
    type: object
  internal_handlers.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_handlers.ErrorDetail'
    type: object
...
  internal_handlers.FraudMetricsResponse:
    properties:
      blocked_clicks:
        example: 0
        type: integer
      blocked_impressions:
        example: 0
        type: integer
      fraud_attempts_count:
        example: 0
        type: integer
      fraud_score_distribution:
        $ref: '#/definitions/internal_handlers.FraudScoreDistribution'
    type: object
  internal_handlers.FraudScoreDistribution:
    properties:
      high:
        example: 0
        type: integer
      low:
        example: 0
        type: integer
      medium:
        example: 0
        type: integer
    type: object
...
  internal_handlers.RevenueAnalyticsResponse:
    properties:
      average_cpc_cents:
        type: number
      average_cpm_cents:
        type: number
      ctr:
        type: number
      from:
        type: string
      to:
        type: string
      total_clicks:
        type: integer
      total_impressions:
        type: integer
      total_revenue_cents:
        type: integer
    type: object
...
  internal_handlers.SimpleErrorResponse:
    properties:
      error:
        example: authentication required
        type: string
    type: object
...
  internal_handlers.TopAdsResponse:
    properties:
      count:
        example: 0
        type: integer
      metric:
        example: impressions
        type: string
      top_ads:
        items:
          additionalProperties: true
          type: object
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@radiobackend.com
    name: API Support
  description: "Radio streaming API with JWT authentication, analytics and multi-language
    support (i18n).\n\n**Security Features (v2.1):**\n- \U0001F510 Timing attack prevention
    (constant-time operations)\n- \U0001F512 HTTPS enforcement (production)\n- \U0001F6AB
    Account lockout (10 attempts = 30 min lockout)\n- \U0001F4E7 Email rate limiting
    (10 attempts/hour per email)\n- \U0001F511 Password strength enforcement (special
    chars + 47 common passwords blocked)\n- \U0001F6E1️ Session hijacking protection
    (User-Agent validation)\n- \U0001F4DD Security event logging (failed attempts
    with metadata)\n- ⚡ Rate limiting (5 attempts/15min per IP)\n- \U0001F510 JWT
    with RS256 + token revocation\n- \U0001F916 Bot protection via shared secret (X-Rradio-Secret
    header)\n\n**Internationalization:**\nSupported languages: Spanish (es), English
    (en), French (fr), German (de).\nUse the 'lang' parameter or 'Accept-Language'
    header to specify the desired language.\n\n**Security Score:** 100/100 (audited
    Dec 2025)"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Radio Backend API
  version: "2.1"
paths:
...
  /api/v1/admin/analytics/campaigns:
    get:
      description: Returns performance metrics (impressions, clicks, CTR, spend) for
        all active campaigns.
      produces:
      - application/json
      responses:
        "200":
          description: List of campaign performance metrics
          schema:
            $ref: '#/definitions/internal_handlers.CampaignPerformanceListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Failed to retrieve campaigns
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get campaign performance
      tags:
      - Admin Analytics
  /api/v1/admin/analytics/dashboard:
    get:
      description: Returns a summary of active campaigns, total budget, total spend,
        impressions, clicks, CTR, and budget utilization percentage.
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard overview metrics
          schema:
            $ref: '#/definitions/internal_handlers.DashboardOverviewResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Failed to retrieve dashboard data
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get admin dashboard overview
      tags:
      - Admin Analytics
  /api/v1/admin/analytics/fraud:
    get:
      description: Returns fraud detection metrics including blocked impressions,
        blocked clicks, and fraud score distribution.
      produces:
      - application/json
      responses:
        "200":
          description: Fraud detection metrics
          schema:
            $ref: '#/definitions/internal_handlers.FraudMetricsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
      security:
      - BearerAuth: []
      summary: Get fraud metrics
      tags:
      - Admin Analytics
  /api/v1/admin/analytics/revenue:
    get:
      description: Returns aggregated revenue metrics (impressions, clicks, CPM, CPC,
        CTR) for a given date range. Defaults to the last 30 days.
      parameters:
      - description: Start date (RFC3339)
        example: '"2024-01-01T00:00:00Z"'
        in: query
        name: from
        type: string
      - description: End date (RFC3339)
        example: '"2024-01-31T23:59:59Z"'
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Revenue analytics data
          schema:
            $ref: '#/definitions/internal_handlers.RevenueAnalyticsResponse'
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
      security:
      - BearerAuth: []
      summary: Get revenue analytics
      tags:
      - Admin Analytics
  /api/v1/admin/analytics/top-ads:
    get:
      description: Returns the top performing advertisements ranked by the specified
        metric (impressions, clicks, ctr, revenue).
      parameters:
      - default: impressions
        description: 'Ranking metric: impressions | clicks | ctr | revenue'
        in: query
        name: metric
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Top ads list with the applied metric
          schema:
            $ref: '#/definitions/internal_handlers.TopAdsResponse'
        "400":
          description: Invalid metric value
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
      security:
      - BearerAuth: []
      summary: Get top performing ads
      tags:
      - Admin Analytics
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  SharedSecret:
    description: Shared secret key required for public station and SEO endpoints (bot
      protection).
    in: header
    name: X-Rradio-Secret
    type: apiKey
swagger: "2.0"